initial class definition.
Initial definition of the abstract Audit base class. Currently the class
provides 4 attributes associated with the general audit structure:
alpha i.e. risk-limit, beta, a maximum number of ballots to draw during
the audit, and replacement. The init method currently only accepts
alpha, but could be expanded to accept others. It is not extremely
important how this is implemented as it should only be called within a
subclass's __init__. There is one abstract method for computing risk to
be implemented within unique audit subclass's. We still need to determine how
the audit objects should accept election information.