initial draft of a contest class. updated audit and brla to incorporate.
Initial draft of a Contest object simply to see how it might
interact with the audit class and subclasses. The ContestType class
could be used to store different types of contests or might not be
necessary.  It has been incorportaed into the Audit init and
subsequently BRLA has been updated. This is all very rough draft, naming
conventions can/will change.