Made final tweaks on PR.
Updated TODO, FIXME sentence. Added 50 char limit to commit message
instructions. The style guide is entirely based on PEP8 and does not
modify it (other than the line length) but simply clarifies in places
where PEP8 is not specific. The PEP8 documents are public domain and it
does not feel necessary to add the citation everywhere, if we want to
add links to headers within the PEP8 page we could later.