added method to compute sample sizes.
Added an abstract method to compute sample sizes to the general Audit
class. This can add functionality similar to what Arlo has, but might be
implemented differently for different audits. For example, it could be
used in BRLA to generate a round shcedule for a given number of rounds
before the audit begins.