tests added for Audit __init__ method.
Audit __init__ method now raises type and value errors for the
parameters entered. Currently, Audit allows the contest input to be None
to allow for testing of the Audit methods independent of the Contest
implementation, but this could be changed later. A test_audit file has
been added with a Simple subclass of Audit just for testing the Audit
class (because it is an ABC). There are currently two tests: 1.
Initialization storing paramteres correctly, and dumby methods returning
correctly, 2. Errors being raised correctly.