Added methods to parse contest, election from json
Added methods for parsing a list of contests or an election from JSON
file. These should make it easier to design and implement tests.
Currently the contest method will produce a list of contests using the
format given in contest_template.json and the election method will parse
a single election (containing a list of contests) using the format given
in election_template.json. There is no real edge case testing for either
of these methods, not sure how much of this needs to be added since they
should only be used for tests developers write.