small fixes. additional tests.
small fixes to current dist functions to test for exact number of rounds
and append sum of distribution sections to risk and stopping probability
schedules. caught and fixed some rounding errors in brla.py along with
some reformatting. added very simple tests to audit and brla for
distribution functions and some of the other new audit functions. some
tests in brla are placeholders for the true tests of the distribution
functions when they are implemented for audits without replacement, for
now simply test that they are not changed by those functions. Also, the
generate contests method in util now makes contests of a fixed size as
writing test cases for variable size contests is strange. Will update
the util methods later.