initial cli implementation.
Summary of changes:
- cli.py contains initial cli implementation of R2B2. Most work is done
with the interactive option, eventually we can add a 'bulk auditing'
option. The option to run in election mode has been added but not
implemented. Main implementation is an interactive audit where the
contest and audit can either be created interactively by the user of
given as a set of cli arguments.
- src/r2b2/tests/data directory was created to store .json and .txt
files that hold data used within the test files and/or templates.
- test_cli contains 4 basic test for what is currently implemented in
cli and uses the expected output files in data/
- function added to util.py to parse a single contest instead of a list
and is used in cli. a template for the single contest JSON format has
been added as well.