better autoapi configuration
Changed autoapi configuration to improve documentation appearance.
Updated the src directory, moved the API reference to the top of the RTD
index, removed members without docstrings from documentation. Added
docstring to r2b2/__init__.py so that r2b2 module will be included in
the documentation. To include other modules, docstrings must be added,
currently r2b2/test is not included, as I am not sure if our test code
belongs in the API documentation or in a more fitting location. Moved
the relevant parts of overview.rst into index and deleted the overview
file alltogether as it is no longer necessary. This setup gives a much
better home page.