added progress bar with verbose bulk option
both bulk methods now include a progress parameter that defaults to
false. if set to true and implemented in the audit a progress bar will
display. cli sets progress true in bulk command with -v flag. can also
be used in any script by setting the parameter to true. Especially
useful for generating large sets of audit data. It indicates to the user
that the computation is, in fact, progressing and estimated remaining
time.