basic simulation interface.
added simulator.py which contains methods for simple mongo db operations
and an abstract Simulation class definition. The idea is that
simulations will be built off of Simulation with the same, very basic
structure, and can be customized for specific cases. Ideally, when
creating a new simulation the developer will not need to do/know much
about the database operations and the functionality they need will be
provided in simulator.py.