updated simulations for database or local mode.
contest.py updated for Contest to have a to_json method that is
currently for local simulation output, but could be helpful in other
parts of the library. simulator.py has been updated to separate database
operations into a class instead of global methods. As it should be
possible for someone to use R2B2 locally to run their own simulations
without our database (or the use of a database ate all), there cannot be
a global MongoClient. Thus, the MongoClient() function only gets called
when database mode is used, i.e. it will not cause unjust exceptions for
those not running/connecting to a database. simulation objects now
maintain a list of trials (as trial entry dicts) created by run() while
the simultion object exists. In the future, there may be a way to load a
locally run simulation object from a saved state, but this is not
important right now. The output method is made available for running
locally to either print out the formatted JSON data for the simulation
or write it to a file. output_audit is separated so that it may be
overwritten by specific simulations with more per-audit information than
the default.