created PairwiseContest class
created pairwise contest class to give the sub-contests better
structure. it holds basic information and does no real error checking
because it should only be accessed from within contest. the single
contest template and all test files that use it needed to be updated to
contain a valid pair because the template contained a tied election,
which would not contain any valid pairwise contests.