audits updated to use pairwise contest
audits now accept a contest and a pair (from which the pairwise contest
is derived) upon instantiation. if no pair is specified, the default is
the top 2 candidates. audit methods now use the pairwise ballot total
instead of the contest total to eliminate irrelevant ballots. brla and
minerva have been updated to reflect the constructor accepting a pair
and minerva uses the pairwise contest info instead of the full contest
info.