updates to analysis of sims.
added method to add an analysis entry to simulations by sim_id. the
minerva simulation has been updated so that analysis only prints data if
verbose and only displays histograms if indicated. It now returns the
experimental risk given the trials and updates it in the analysis entry
of the simulation using the new db method. some import fixes and
reformatting as well.