db authentication added.
the r2b2 mongoDB has user authentication to read/write to the database.
the default is set to reader in the simulator module so a contributor
must knowingly modify to write to be able to modify the database.
additionally scripts could be written to use personal accounts instead
of the reader write accounts.

also added pymongo and matplotlib to setup requirements.