initial multiround mineva sim added
Added class for a multiround Minerva risk simulation. The class is
simply a draft of the simulation and it untested. The format of the
output data from trial needs to be corrected. the analyze method needs
to be finished. the simulation must be tested before it is used to
generate real data.