Fixed rare bug (numerical imprecision occurring in very large distributions)
- By checking the stopping condition as a weak inequality was causing a problem because
sometimes (for large distributions only) both the numerator and denominator were essentially 0
and this was being interpretted as not meeting the risk, but in fact these cases were occuring
at the far right end of the distribution and thus should have been passing the check.

- This bug is relatively complicated and not worth explaining in this commit message in full depth,
but contact Oliver Broadrick for more info/questions.